Restructuring -yaml not installed

+++ InventoryManagement/IMS2/db/db_utils.py
@@ -1,4 +1,3 @@
-import os
 import asyncio
 import asyncpg
 from asyncpg import UndefinedTableError
@@ -7,36 +6,59 @@ from types import TracebackType
 from typing import Optional, Type, List, Tuple
 from common.d_logger import Logs
 from constants import ConfigReader
+import logging
+
+logger = Logs().get_logger("db")
+
+
+async def connect_pg(db_config_file):
+    config = ConfigReader(db_config_file)
+    try:
+        conn = await asyncpg.connect(host=config.get_options("Host"),
+                                     port=config.get_options("Port"),
+                                     user=config.get_options("User"),
+                                     database=config.get_options("Database"),
+                                     password=config.get_options("Password"))
+        return conn
+    except Exception as e:
+        logger.exception('Error while connecting to DB', e)
+        raise e
+
 
 
 class ConnectPg:
-    def __init__(self, config):
-        self.config = config
+    def __init__(self, db_config_file):
+        self.config = ConfigReader(db_config_file)
         self._conn = None
 
     async def __aenter__(self):
+        logger.debug("Trying to connect to db ...")
+        logger.debug("Entering context manager, waiting for connection")
         try:
             self._conn = await asyncpg.connect(host=self.config.get_options("Host"),
                                                port=self.config.get_options("Port"),
                                                user=self.config.get_options("User"),
                                                database=self.config.get_options("Database"),
                                                password=self.config.get_options("Password"))
+            logger.debug("Successfully connected!!!")
             return self._conn
         except Exception as e:
-            raise e
+            logger.exception('Error while connecting to DB', e)
+            return None
 
     async def __aexit__(self,
                         exc_type: Optional[Type[BaseException]],
                         exc_val: Optional[BaseException],
                         exc_tb: Optional[TracebackType]):
+        logger.debug("Exiting context manager")
         if self._conn:
+            logger.debug("Closed connection")
             await self._conn.close()
 
 
 class DbUtil:
-    def __init__(self):
-        self.logger = Logs().get_logger(os.path.basename(__file__))
-        self.config = ConfigReader()
+    def __init__(self, db_config_file: str):
+        self.db_config_file = db_config_file
 
     async def create_tables(self, statements: List[str]):
         """
@@ -45,22 +67,22 @@ class DbUtil:
         :param statements: sql statments
         :return:
         """
-        self.logger.debug("Creating tables")
         results = []
-        try:
-            async with ConnectPg(self.config) as conn:
-                for statement in statements:
-                    try:
-                        self.logger.info(f"{statement}")
-                        status = await conn.execute(statement)
-                        results.append(status)
-                        self.logger.debug(status)
-                    except Exception as e:
-                        self.logger.error(f'create_tables: Error while creating table: {statement}', e)
-                self.logger.info("Tables created")
-        except Exception as e:
-            self.logger.error(e)
+        async with ConnectPg(self.db_config_file) as conn:
+            if conn is None:
+                logger.debug("Error while connecting to DB during creating tables")
+                return
 
+            logger.info("Creating the tables")
+            for statement in statements:
+                try:
+                    logger.info(f"{statement}")
+                    status = await conn.execute(statement)
+                    results.append(status)
+                    logger.debug(status)
+                except Exception as e:
+                    logger.exception(f'create_tables: Error while creating table: {statement}', e)
+            logger.info("Finished creating the tables")
         return results
 
     async def drop_tables(self, table_names: List[str]):
@@ -70,23 +92,23 @@ class DbUtil:
         :return:  the list of results of dropping the tables or
                   None if connection fails
         """
-        self.logger.debug(f"Dropping tables {table_names}")
         results = []
-        try:
-            async with ConnectPg(self.config) as conn:
-                for table in table_names:
-                    try:
-                        sql_stmt = f'DROP TABLE {table} CASCADE;'
-                        result = await conn.execute(sql_stmt)
-                        results.append(result)
-                    except UndefinedTableError as ute:
-                        self.logger.error('drop_table: Trying to drop an undefined table', ute)
-                    except Exception as e:
-                        self.logger.error('drop_table: Error while dropping tables', e)
-            self.logger.info("Tables removed")
-        except Exception as e:
-            self.logger.error(e)
+        async with ConnectPg(self.db_config_file) as conn:
+            if conn is None:
+                logger.debug("Error while connecting to DB during removing tables")
+                return None
 
+            logger.info("Removing the tables")
+            for table in table_names:
+                try:
+                    sql_stmt = f'DROP TABLE {table} CASCADE;'
+                    result = await conn.execute(sql_stmt)
+                    results.append(result)
+                except UndefinedTableError as ute:
+                    logger.exception('drop_table: Trying to drop an undefined table', ute)
+                except Exception as e:
+                    logger.exception('drop_table: Error while dropping tables', e)
+        logger.info("Finished removing the tables")
         return results
 
     async def select_query(self, query: str, args: List = None):
@@ -95,21 +117,22 @@ class DbUtil:
         :param query
         :return: all results if successful, otherwise None
         """
-        results = []
-        try:
-            async with ConnectPg(self.config) as conn:
-                try:
-                    query = await conn.prepare(query)
-                    if args:
-                        results: List[Record] = await query.fetch(*args)
-                    else:
-                        results: List[Record] = await query.fetch()
-                except Exception as e:
-                    self.logger.error(f'select_query: Error while executing {query}', e)
-        except Exception as e:
-            self.logger.error(e)
+        async with ConnectPg(self.db_config_file) as conn:
+            if conn is None:
+                logger.debug("Error while connecting to DB during querying tables")
+                return None
 
-        return results
+            try:
+                query = await conn.prepare(query)
+                if args:
+                    results: List[Record] = await query.fetch(*args)
+                else:
+                    results: List[Record] = await query.fetch()
+                return results
+            except Exception as e:
+                logger.exception(f'select_query: Error while executing {query}', e)
+                return None
+
 
     async def executemany(self, statement: str, args: List[Tuple]):
         """
@@ -120,22 +143,20 @@ class DbUtil:
             if successful, None
             otherwise, exception or string
         """
-        results = None
-        self.logger.debug(f"executemany({statement}, {args})")
+        async with ConnectPg(self.db_config_file) as conn:
+            if conn is None:
+                logger.debug("Error while connecting to DB during sync_executing")
+                return "Connection failed"
 
-        try:
-            async with ConnectPg(self.config) as conn:
-                self.logger.debug("Synchronous executing")
-                try:
-                    results = await conn.executemany(statement, args)
-                    self.logger.debug(f"results::\n{results}")
-                except Exception as e:
-                    results = e
-                    self.logger.error('executemany: Failed synchronous executing')
-        except Exception as e:
-            self.logger.error(e)
+            logger.info("Synchronous executing")
+            try:
+                results = await conn.executemany(statement, args)
+                logger.info(f"results::\n{results}")
+                return results
+            except Exception as e:
+                logger.exception('executemany: Error during synchronous executing', e)
+                return e
 
-        return results
 
     async def pool_execute(self, statement: str, args: List[Tuple]):
         """
@@ -148,25 +169,21 @@ class DbUtil:
         """
         async def execute(stmt, arg, pool):
             async with pool.acquire() as conn:
-                self.logger.debug(stmt)
-                self.logger.debug(arg)
+                logger.debug(stmt)
+                logger.debug(arg)
                 return await conn.execute(stmt, *arg)
 
-        self.logger.debug("Asynchronous executing")
-        results = []
-        try:
-            async with asyncpg.create_pool(host=self.config.get_options("Host"),
-                                           port=self.config.get_options("Port"),
-                                           user=self.config.get_options("User"),
-                                           database=self.config.get_options("Database"),
-                                           password=self.config.get_options("Password")) as pool:
-                queries = [execute(statement, arg, pool) for arg in args]
-                results = await asyncio.gather(*queries, return_exceptions=True)
-                self.logger.debug(f":\n{results}")
-        except Exception as e:
-            self.logger.error(e)
-
-        return results
+        logger.info("Asynchronous executing")
+        config = ConfigReader(self.db_config_file)
+        async with asyncpg.create_pool(host=config.get_options("Host"),
+                                       port=config.get_options("Port"),
+                                       user=config.get_options("User"),
+                                       database=config.get_options("Database"),
+                                       password=config.get_options("Password")) as pool:
+            queries = [execute(statement, arg, pool) for arg in args]
+            results = await asyncio.gather(*queries, return_exceptions=True)
+            logger.debug(f":\n{results}")
+            return results
 
     async def delete(self, table, col_name, args: List[Tuple]):
         """
@@ -183,17 +200,18 @@ class DbUtil:
                 otherwise, exception
         """
         if not isinstance(args, List):
-            self.logger.error(f"args' type{type(args)} must be List[Tuple]")
+            logger.error(f"args' type{type(args)} must be List[Tuple]")
             return None
         if not isinstance(args[0], Tuple):
-            self.logger.error(f"args element's type{type(args[0])} must be Tuple")
+            logger.error(f"args element's type{type(args[0])} must be Tuple")
             return None
 
         stmt = f"DELETE FROM {table} WHERE {col_name} = $1"
 
-        self.logger.debug(f"Delete rows ...")
-        self.logger.debug(args)
+        logger.debug(f"Delete rows ...")
+        logger.debug(args)
 
+        # results = await self.pool_execute(stmt, args)
         results = await self.executemany(stmt, args)
-        self.logger.debug(f":\n{results}")
+        logger.debug(f":\n{results}")
         return results
